refactor: 회원의 repository 패키지를 domain 안으로 이동
